Auto is back on Track, tuning opMode created. ALl that needs to be done is testing lambda commands and actions.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Kotlin_Bromine_Arya/Tele/SubSystems/Shoulder.kt
@@ -57,7 +57,7 @@ open class Shoulder(hwMap: HardwareMap) {
     //run consistantly
     fun pidShoulder() {
         findTarget()
-        val pidf = shoulderPIDF.calculate(shoulder.currentPosition.toDouble(), target.toDouble())
+        val pidf = shoulderPIDF.calculate(shoulder.currentPosition.toDouble(), target)
         shoulder.power =  if (target == 0.0) { powerDrop } else { pidf }
     }
